[browsable] extract RWrapper, RGroup, RProvider from RBrowsable.hxx
