+++ gui/browsable/src/RBrowsable.cxx
@@ -21,187 +21,6 @@ using namespace ROOT::Experimental::Browsable;
 using namespace std::string_literals;
 
 
-class RCompositeIter : public RLevelIter {
-   int fIndx{-1};
-   RComposite &fComp;
-
-public:
-
-   explicit RCompositeIter(RComposite &comp) : fComp(comp) {}
-   virtual ~RCompositeIter() = default;
-
-   /** Shift to next element */
-   bool Next() override { fIndx++; return HasItem(); }
-
-   /** Is there current element  */
-   bool HasItem() const override { return (fIndx >= 0) &&  (fIndx < (int) fComp.GetChilds().size()); }
-
-   /** Returns current element name  */
-   std::string GetName() const override { return fComp.GetChilds()[fIndx]->GetName(); }
-
-   /** If element may have childs: 0 - no, >0 - yes, -1 - maybe */
-   int CanHaveChilds() const override { return fComp.GetChilds().size(); }
-
-   /** Returns full information for current element */
-   std::shared_ptr<RElement> GetElement() override { return fComp.GetChilds()[fIndx]; }
-
-   /** Reset iterator to the first element, returns false if not supported */
-   bool Reset() override { fIndx = -1; return true; }
-
-   /** Find item with specified name, use item MatchName() functionality */
-   bool Find(const std::string &name) override
-   {
-      if (!Reset()) return false;
-
-      while (Next()) {
-         if (fComp.GetChilds()[fIndx]->MatchName(name))
-            return true;
-      }
-
-      return false;
-   }
-
-};
-
-/////////////////////////////////////////////////////////////////////
-/// Create iterator for childs of composite
-
-std::unique_ptr<RLevelIter> RComposite::GetChildsIter()
-{
-   return std::make_unique<RCompositeIter>(*this);
-}
-
-/////////////////////////////////////////////////////////////////////
-/// Return icon name for the given class
-/// TODO: should be factorized out from here
-
-std::string RProvider::GetClassIcon(const std::string &classname)
-{
-   if (classname == "TTree" || classname == "TNtuple")
-      return "sap-icon://tree"s;
-   if (classname == "TDirectory" || classname == "TDirectoryFile")
-      return "sap-icon://folder-blank"s;
-   if (classname.find("TLeaf") == 0)
-      return "sap-icon://e-care"s;
-
-   return "sap-icon://electronic-medical-record"s;
-}
-
-
-RProvider::BrowseMap_t &RProvider::GetBrowseMap()
-{
-   static RProvider::BrowseMap_t sMap;
-   return sMap;
-}
-
-RProvider::FileMap_t &RProvider::GetFileMap()
-{
-   static RProvider::FileMap_t sMap;
-   return sMap;
-}
-
-
-RProvider::~RProvider()
-{
-   // here to remove all correspondent entries
-   auto &fmap = GetFileMap();
-
-   for (auto fiter = fmap.begin();fiter != fmap.end();) {
-      if (fiter->second.provider == this)
-         fiter = fmap.erase(fiter);
-      else
-         fiter++;
-   }
-
-   auto &bmap = GetBrowseMap();
-   for (auto biter = bmap.begin(); biter != bmap.end();) {
-      if (biter->second.provider == this)
-         biter = bmap.erase(biter);
-      else
-         biter++;
-   }
-}
-
-//////////////////////////////////////////////////////////////////////////////////
-// Register file open function for specified extension
-
-void RProvider::RegisterFile(const std::string &extension, FileFunc_t func)
-{
-    auto &fmap = GetFileMap();
-
-    if ((extension != "*") && (fmap.find(extension) != fmap.end()))
-       R__ERROR_HERE("Browserv7") << "Provider for file extension  " << extension << " already exists";
-
-    fmap.emplace(extension, StructFile{this,func});
-}
-
-//////////////////////////////////////////////////////////////////////////////////
-// Register browse function for specified class
-
-void RProvider::RegisterBrowse(const TClass *cl, BrowseFunc_t func)
-{
-    auto &bmap = GetBrowseMap();
-
-    if (cl && (bmap.find(cl) != bmap.end()))
-       R__ERROR_HERE("Browserv7") << "Browse provider for class " << cl->GetName() << " already exists";
-
-    bmap.emplace(cl, StructBrowse{this,func});
-}
-
-//////////////////////////////////////////////////////////////////////////////////
-// remove provider from all registered lists
-
-std::shared_ptr<RElement> RProvider::OpenFile(const std::string &extension, const std::string &fullname)
-{
-   auto &fmap = GetFileMap();
-
-   auto iter = fmap.find(extension);
-
-   if (iter != fmap.end()) {
-      auto res = iter->second.func(fullname);
-      if (res) return res;
-   }
-
-   for (auto &pair : fmap)
-      if ((pair.first == "*") || (pair.first == extension)) {
-         auto res = pair.second.func(fullname);
-         if (res) return res;
-      }
-
-   return nullptr;
-}
-
-/////////////////////////////////////////////////////////////////////////
-/// Create browsable element for the object
-/// Created element may take ownership over the object
-
-std::shared_ptr<RElement> RProvider::Browse(std::unique_ptr<Browsable::RHolder> &object)
-{
-   if (!object)
-      return nullptr;
-
-   auto &bmap = GetBrowseMap();
-
-   auto cl = object->GetClass();
-   if (!cl)
-      return nullptr;
-
-   auto iter = bmap.find(cl);
-
-   if (iter != bmap.end()) {
-      auto res = iter->second.func(object);
-      if (res || !object) return res;
-   }
-
-   for (auto &pair : bmap)
-      if ((pair.first == nullptr) || (cl == pair.first)) {
-         auto res = pair.second.func(object);
-         if (res || !object) return res;
-      }
-
-   return nullptr;
-}
-
 /////////////////////////////////////////////////////////////////////
 /// set top element for browsing
